refactor: extract logic for ChatArea component into custom hook. Remove unused imports
