+++ src/shared/ui/scroll/Scroll.tsx
@@ -1,14 +1,14 @@
 import * as ScrollArea from '@radix-ui/react-scroll-area'
-import s from './scroll.module.scss'
 import clsx from 'clsx'
-import { ComponentPropsWithoutRef, ElementRef, forwardRef } from 'react'
+import { ComponentPropsWithoutRef, forwardRef } from 'react'
+import s from './scroll.module.scss'
 
 type Props = {
   children: React.ReactNode
   className?: string
 } & ComponentPropsWithoutRef<typeof ScrollArea.Root>
 
-export const Scroll = forwardRef<HTMLDivElement, Props>(({ children, className, ...rest }, ref) => {
+export const Scroll = forwardRef<HTMLDivElement, Props>(({ children, className }, ref) => {
   return (
     <ScrollArea.Root className={clsx(s.root, className)}>
       <ScrollArea.Viewport className={s.viewport} ref={ref}>
